+++ main.rb
@@ -1,7 +1,9 @@
+# frozen_string_literal: true
+
 require_relative './lib/rest_client_test.rb'
 
-searching = RestClientSearch.new()
-puts "Enter your search query"
+searching = RestClientSearch.new
+puts 'Enter your search query'
 searching.search_for?
 p 'Which website to search Google or Bing?'
 searching.website_to_search?
@@ -12,6 +14,3 @@ searching.rest_get_parsed
 searching.parse_it
 searching.print_links
 p searching.array_links
-
-
-
